Move ligth direction param item to params package
